Update conflict constraint for "nelmio/api-doc-bundle" in order to allow version 3
Update src/Controller/Api/Legacy/GroupController.php


Update src/Controller/Api/Legacy/UserController.php


Update tests/Controller/Api/GroupControllerTest.php


Update tests/Controller/Api/UserControllerTest.php
