+++ UPGRADE-4.x.md
@@ -1,6 +1,13 @@
 UPGRADE 4.x
 ===========
 
+UPGRADE FROM 3.x to 3.x
+=======================
+
+### Support for NelmioApiDocBundle > 3.6 is added
+
+Controllers for NelmioApiDocBundle v2 were moved under `Sonata\UserBundle\Controller\Api\Legacy\` namespace and controllers for NelmioApiDocBundle v3 were added as replacement. If you extend them, you must ensure they are using the corresponding inheritance.
+
 UPGRADE FROM 4.6 to 4.7
 ========================
