+++ composer.json
@@ -44,7 +44,7 @@
     "conflict": {
         "friendsofsymfony/rest-bundle": "<2.1",
         "jms/serializer": "<1.3 || >=4.0",
-        "nelmio/api-doc-bundle": "<2.4",
+        "nelmio/api-doc-bundle": "<2.13 || >=4.0",
         "sonata-project/core-bundle": "<3.20",
         "sonata-project/doctrine-orm-admin-bundle": "<3.0",
         "sonata-project/google-authenticator": "<1.0"
@@ -56,11 +56,12 @@
         "jms/serializer-bundle": "^1.0 || ^2.0 || ^3.0",
         "matthiasnoback/symfony-config-test": "^4.0",
         "matthiasnoback/symfony-dependency-injection-test": "^4.0",
-        "nelmio/api-doc-bundle": "^2.4",
+        "nelmio/api-doc-bundle": "^2.13.4 || ^3.6",
         "sonata-project/doctrine-orm-admin-bundle": "^3.18",
         "sonata-project/google-authenticator": "^1.0 || ^2.0",
         "symfony/browser-kit": "^4.4 || ^5.1",
-        "symfony/phpunit-bridge": "^5.1.1"
+        "symfony/phpunit-bridge": "^5.1.1",
+        "symfony/swiftmailer-bundle": "^3.4"
     },
     "suggest": {
         "friendsofsymfony/rest-bundle": "For using the public API methods.",
