+++ src/DependencyInjection/SonataUserExtension.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sonata\UserBundle\DependencyInjection;
 
+use Nelmio\ApiDocBundle\Annotation\Operation;
 use Sonata\Doctrine\Mapper\Builder\OptionsBuilder;
 use Sonata\Doctrine\Mapper\DoctrineCollector;
 use Sonata\EasyExtendsBundle\Mapper\DoctrineCollector as DeprecatedDoctrineCollector;
@@ -93,7 +94,11 @@ class SonataUserExtension extends Extension implements PrependExtensionInterface
             $loader->load('serializer.xml');
 
             $loader->load('api_form.xml');
-            $loader->load('api_controllers.xml');
+            if (class_exists(Operation::class)) {
+                $loader->load('api_controllers.xml');
+            } else {
+                $loader->load('api_controllers_legacy.xml');
+            }
         }
 
         if ($config['security_acl']) {
