+++ tests/App/Resources/config/routing/api_nelmio_v3.yml
@@ -1,0 +1,13 @@
+app.swagger:
+    path: /api/doc.json
+    methods: GET
+    defaults: {_controller: nelmio_api_doc.controller.swagger}
+
+app.swagger_ui:
+    path: /api/doc
+    methods: GET
+    defaults: {_controller: nelmio_api_doc.controller.swagger_ui}
+
+sonata_api_user:
+    prefix: /api/user
+    resource: "@SonataUserBundle/Resources/config/routing/api_nelmio_v3.xml"
