+++ tests/Functional/Routing/RoutingTest.php
@@ -13,7 +13,8 @@ declare(strict_types=1);
 
 namespace Sonata\UserBundle\Tests\Functional\Routing;
 
-use Sonata\UserBundle\Tests\Functional\App\AppKernel;
+use Nelmio\ApiDocBundle\Annotation\Operation;
+use Sonata\UserBundle\Tests\App\AppKernel;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 /**
@@ -39,7 +40,7 @@ final class RoutingTest extends WebTestCase
 
         $matchingPath = $path;
         $matchingFormat = '';
-        if (false !== strpos($matchingPath, '.{_format}', -10)) {
+        if (\strlen($matchingPath) >= 10 && false !== strpos($matchingPath, '.{_format}', -10)) {
             $matchingFormat = '.json';
             $matchingPath = str_replace('.{_format}', $matchingFormat, $path);
         }
@@ -74,7 +75,15 @@ final class RoutingTest extends WebTestCase
 
     public function getRoutes(): iterable
     {
-        yield ['nelmio_api_doc_index', '/api/doc/{view}', ['GET']];
+        // API
+        if (class_exists(Operation::class)) {
+            yield ['app.swagger_ui', '/api/doc', ['GET']];
+            yield ['app.swagger', '/api/doc.json', ['GET']];
+        } else {
+            yield ['nelmio_api_doc_index', '/api/doc/{view}', ['GET']];
+        }
+
+        // API - User
         yield ['sonata_api_user_user_get_users', '/api/user/users.{_format}', ['GET']];
         yield ['sonata_api_user_user_get_user', '/api/user/user/{id}.{_format}', ['GET']];
         yield ['sonata_api_user_user_post_user', '/api/user/user.{_format}', ['POST']];
@@ -82,6 +91,8 @@ final class RoutingTest extends WebTestCase
         yield ['sonata_api_user_user_delete_user', '/api/user/user/{id}.{_format}', ['DELETE']];
         yield ['sonata_api_user_user_post_user_group', '/api/user/user/{userId}/{groupId}.{_format}', ['POST']];
         yield ['sonata_api_user_user_delete_user_group', '/api/user/user/{userId}/{groupId}.{_format}', ['DELETE']];
+
+        // API - Group
         yield ['sonata_api_user_group_get_groups', '/api/user/groups.{_format}', ['GET']];
         yield ['sonata_api_user_group_get_group', '/api/user/group/{id}.{_format}', ['GET']];
         yield ['sonata_api_user_group_post_group', '/api/user/group.{_format}', ['POST']];
